refactor(details): `arithmetic_operator` -> `binary_fn` + `applicable_on_tuples`
The class `arithmetic_operator` was representing an associative operator accepting two arguments via `operator()` or a binary tuple via the special member `tuple`. Not only it was very limited in use (only associative operators? It is ok for a binary function, but that's all), it was implemented recursively
so it was not compile-time-friendly.
In this commit the `arithmetic_operator` splits in two:
- the already existing `binary_fn`
- a generic `applicable_on_tuples` accepting an N-ary tuple an applying
  `Base{}` to all the arguments at the same time.
Decoupling `applicable_on_tuples` allows me to use it in non-binary
functions